Added basic target manager
The target manager can hold a list of targets. The targets are told to
be added by the Level class calling an instance of the AddBullets()
method inside the target manager. When this is called it is passed a
position which is then used when calling the setup() for the target to
determine each targets position.