Added new Recoil type
Recoil no longer recoils the crosshair to the next position you are
going to shoot. Instead it recoils straight up and back down each time
the player shoots.

The recoil in this type is where the bullets land.
There is now a recoil cool down time, if the player shoots within this
time since their last shot, the bullets will recoil more. If the player
consecutively shoots within the cool down time, the cool down timer is
restarted. The recoil for the y direction is also multiplied.

If the player lets the recoil timer cool down to zero, all recoil
variables are reset to normal. Allowing them to shoot accurately again.

The faster the player shoots, the more they will be punished, in terms
of recoil.

This recoil type requires the player to be more skilful.

If the player gets good enough, they can control the recoil and shoot
accurately even when the recoil is being multiplied.