Added a reload button
The reload button has a normal reload state and quick reload state.

The quick reload state is triggered when the players clip is fully
depleted. It then has a timer, if the player clicks reload during this
time (indicated by sprites) the gun will instantly reload.

If the player clicks reload when the quick reload state is not active,
reloading will take a certain time.

Separate sounds will play for the quick reload and normal reload.