Changes
Each bullet is now assigned the number target it hit.
I will try to use this to decide which bullets to not draw when I stop
drawing a target because it has been shot enough times to be taken down.

Inside the target collision I no longer create a local copy of the list
of targets. I am now passing the list by reference so I can manipulate
the target being shot inside the collision check using the iterator
which points at the current target.