Game Time / Targets Respawn / Adjustments
Bullets that land on the wall no longer are drawn above the targets.

Added right mouse click which changes the player crosshair to a down the
sights type crosshair.

Targets can respawn when the reload button is pressed if there are no
targets left.

Game time is now displayed in the window.