Info pop up edit
Info pop ups are now triggered on and off by the players actions. So if
the player does what the info is telling them to do, it will no longer
display.

Info pop ups won't display while another is displayed.
If it should be triggered while another is displayed, it will wait until
the next time it is triggered and an info is not being displayed
currently, to display the next info pop up.