Added level 2 assets & performance improvement
Level 2 assets added by layer so targets can appear to come out from
behind cover.

Sound Manager is now called before the window is opened so sounds are
loaded before any are played.
This fixes a problem where the sound files would be loaded in and set
when the first sound is played and slow down would occur.