Volume option added
When the options menu element is clicked, the volume option is displayed
and can be interacted with.

Pressing on the arrows increases and decreases the volume.
Arrows move a little when clicked to indicate they are clicked.
They are also highlighted before clicking.

The gunshot noise is played when the button is clicked to let the user
know what the volume they have selected sounds like.

A number from 0-100 is also displayed.