Target Points detection added
When fired upon, targets will display point pop ups as well as adding to
the players overall score. If the target is fired upon soon after the
previous shot, the points will be added to the current points pop up,
and its stall timer will be reset.

When the stall timer runs out the points gained move towards the scores
position. The time to live is started and the points pop up will then
disappear and be deleted from memory.