Recoil modification
Bullets now recoil relative to time passed since the last bullet fired.

When a bullet is fired, a cooldown timer is started/reset.
If a bullet is fired instantly after the previous bullet, it will recoil
at 100% of the recoil strength for that gun.

If a bullet is fired after half of the recoil cooldown time has passed,
the bullet will recoil at 50% of the recoil strength for that gun.

There is also a small recoil multiplier that will increase the strength
of the recoil for that gun each time the gun is fired without letting
the recoil cooldown time pass.