Points multiplier added
If the player shoots at the target, then shoots and hits the same area
as the previous shot within the stall time of the points (0.5 seconds),
a multiplier will be added to the score. This multiplier is determined
by whether the player has hit the same area consecutively on the top
target or bottom target. The top target is smaller so the multiplier
rewarded is larger.

There is also a base multiplier which is applied if the target is hit
anywhere consecutively within the stall time. This multiplier is the
smallest of the 3.