Modified Sound Manager
Changed the sound manager to add sounds with specific buffers to a queue
then plays them. When a sound has stopped playing, it gets deleted. This
enables sounds with the same buffer to be played multiple times in a row
without interrupting the sound playing.

This fixes the gun shells falling sound effect not playing until gun
shots of rapid succession had ceased.